Refactored database models
