--- app/models.py
@@ -1,777 +1,0 @@
-import os
-import cv2
-import json
-import time
-import datetime
-import numpy as np
-import imantics as im
-
-from flask_socketio import emit
-from flask_mongoengine import MongoEngine
-from mongoengine.queryset.visitor import Q
-from flask_login import UserMixin, current_user
-
-from .config import Config
-from PIL import Image
-
-import logging
-logger = logging.getLogger('gunicorn.error')
-
-
-db = MongoEngine()
-
-
-class Event(db.EmbeddedDocument):
-    
-    name = db.StringField()
-    created_at = db.DateTimeField()
-    data = db.GenericEmbeddedDocumentField()
-
-    def set_data(self, event):
-        self.name = event.NAME
-        self.data = event
-        self.created_at = datetime.datetime.now()
-
-
-class DatasetModel(db.DynamicDocument):
-    
-    id = db.SequenceField(primary_key=True)
-    name = db.StringField(required=True, unique=True)
-    directory = db.StringField()
-    thumbnails = db.StringField()
-    categories = db.ListField(default=[])
-
-    owner = db.StringField(required=True)
-    users = db.ListField(default=[])
-
-    annotate_url = db.StringField(default="")
-
-    default_annotation_metadata = db.DictField(default={})
-
-    deleted = db.BooleanField(default=False)
-    deleted_date = db.DateTimeField()
-
-    def save(self, *args, **kwargs):
-
-        directory = os.path.join(Config.DATASET_DIRECTORY, self.name + '/')
-
-        if not os.path.exists(directory):
-            os.makedirs(directory)
-
-        self.directory = directory
-
-        if current_user:
-            self.owner = current_user.username
-        else:
-            self.owner = 'system'
-
-        return super(DatasetModel, self).save(*args, **kwargs)
-    
-    def download_images(self, keywords, limit=100):
-
-        task = TaskModel(
-            name="Downloading {} images to {} with keywords {}".format(limit, self.name, keywords),
-            dataset_id=self.id,
-            group="Downloading Images"
-        )
-
-        def download_images(task, dataset, keywords, limit):
-            def custom_print(string):
-                __builtins__.print("%f -- %s" % (time.time(), string))
-
-                print = dprint
-                task.log()
-            for keyword in args['keywords']:
-                response = gid.googleimagesdownload()
-                response.download({
-                    "keywords": keyword,
-                    "limit": args['limit'],
-                    "output_directory": output_dir,
-                    "no_numbering": True,
-                    "format": "jpg",
-                    "type": "photo",
-                    "print_urls": False,
-                    "print_paths": False,
-                    "print_size": False
-                })
-
-        return task
-
-    def import_coco(self, coco):
-        from .util.task_util import import_coco_func
-        task = TaskModel(
-            name="Import COCO ({})".format(self.name),
-            dataset_id=self.id,
-            group="Annotation Import"
-        )
-        task.save()
-        task.start(import_coco_func, dataset=self, coco_json=coco)
-
-        return task.api_json()
-
-
-    def export_coco(self):
-
-        from .util.task_util import export_coco_func
-        task = TaskModel(
-            name="Export COCO ({})".format(self.name),
-            dataset_id=self.id,
-            group="Annotation Export"
-        )
-        task.save()
-        task.start(export_coco_func, dataset=self)
-
-        return task.api_json()
-
-    def scan(self):
-
-        from .util.task_util import scan_func
-        task = TaskModel(
-            name=f"Scanning {self.name} for new images",
-            dataset_id=self.id,
-            group="Directory Image Scan"
-        )
-        task.save()
-        task.start(scan_func, dataset=self)
-
-        return task.api_json()
-
-    def is_owner(self, user):
-
-        if user.is_admin:
-            return True
-        
-        return user.username.lower() == self.owner.lower()
-
-    def can_download(self, user):
-        return self.is_owner(user)
-
-    def can_delete(self, user):
-        return self.is_owner(user)
-    
-    def can_share(self, user):
-        return self.is_owner(user)
-    
-    def can_generate(self, user):
-        return self.is_owner(user)
-
-    def can_edit(self, user):
-        return user.username in self.users or self.is_owner(user)
-    
-    def permissions(self, user):
-        return {
-            'owner': self.is_owner(user),
-            'edit': self.can_edit(user),
-            'share': self.can_share(user),
-            'generate': self.can_generate(user),
-            'delete': self.can_delete(user),
-            'download': self.can_download(user)
-        }
-
-
-class ImageModel(db.DynamicDocument):
-    
-    # -- Contants
-    THUMBNAIL_DIRECTORY = '.thumbnail'
-    PATTERN = (".gif", ".png", ".jpg", ".jpeg", ".bmp")
-
-    # -- Private
-    _dataset = None
-
-    # -- Database
-    id = db.SequenceField(primary_key=True)
-    dataset_id = db.IntField()
-    category_ids = db.ListField(default=[])
-
-    # Absolute path to image file
-    path = db.StringField(required=True, unique=True)
-    width = db.IntField(required=True)
-    height = db.IntField(required=True)
-    file_name = db.StringField()
-    
-    # True if the image is annotated
-    annotated = db.BooleanField(default=False)
-    # Poeple currently annotation the image
-    annotating = db.ListField(default=[])
-
-    thumbnail_url = db.StringField()
-    image_url = db.StringField()
-    coco_url = db.StringField()
-
-    metadata = db.DictField()
-    license = db.IntField()
-
-    deleted = db.BooleanField(default=False)
-    deleted_date = db.DateTimeField()
-
-    events = db.EmbeddedDocumentListField(Event)
-
-    regenerate_thumbnail = db.BooleanField(default=False)
-
-    @classmethod
-    def create_from_path(cls, path, dataset_id=None):
-
-        pil_image = Image.open(path)
-
-        image = cls()
-        image.file_name = os.path.basename(path)
-        image.path = path
-        image.width = pil_image.size[0]
-        image.height = pil_image.size[1]
-
-        if dataset_id is not None:
-            image.dataset_id = dataset_id
-        else:
-            # Get dataset name from path
-            folders = path.split('/')
-            i = folders.index("datasets")
-            dataset_name = folders[i+1]
-
-            dataset = DatasetModel.objects(name=dataset_name).first()
-            if dataset is not None:
-                image.dataset_id = dataset.id
-
-        pil_image.close()
-
-        return image
-
-    def delete(self, *args, **kwargs):
-        self.thumbnail_delete()
-        AnnotationModel.objects(image_id=self.id).delete()
-        return super(ImageModel, self).delete(*args, **kwargs)
-
-    def thumbnail(self):
-        """
-        Generates (if required) and returns thumbnail
-        """
-        if not self.annotated:
-            self.thumbnail_delete()
-            return Image.open(self.path)
-        
-        thumbnail_path = self.thumbnail_path()
-
-        if self.regenerate_thumbnail or \
-            not os.path.isfile(thumbnail_path):
-            
-            logger.debug(f'Generating thumbnail for {self.id}')
-
-            pil_image = self.generate_thumbnail()
-            pil_image = pil_image.convert("RGB")
-            pil_image.save(thumbnail_path)
-
-            self.update(is_modified=False)
-            return pil_image
-        else:
-            return Image.open(thumbnail_path)
-    
-    def thumbnail_path(self):
-        folders = self.path.split('/')
-        folders.insert(len(folders)-1, self.THUMBNAIL_DIRECTORY)
-
-        path = '/' + os.path.join(*folders)
-        directory = os.path.dirname(path)
-
-        if not os.path.exists(directory):
-            os.makedirs(directory)
-        
-        return path
-    
-    def thumbnail_delete(self):
-        path = self.thumbnail_path()
-        if os.path.isfile(path):
-            os.remove(path)
-
-    def generate_thumbnail(self):
-        image = self().draw(color_by_category=True, bbox=False)
-        return Image.fromarray(image)
-
-    def flag_thumbnail(self, flag=True):
-        """
-        Toggles values to regenerate thumbnail on next thumbnail request
-        """
-        if self.regenerate_thumbnail != flag:
-            self.update(regenerate_thumbnail=flag)
-
-    def copy_annotations(self, annotations):
-        """
-        Creates a copy of the annotations for this image
-        :param annotations: QuerySet of annotation models
-        :return: number of annotations
-        """
-        annotations = annotations.filter(width=self.width, height=self.height, area__gt=0)
-
-        for annotation in annotations:
-            clone = annotation.clone()
-
-            clone.dataset_id = self.dataset_id
-            clone.image_id = self.id
-
-            clone.save(copy=True)
-
-        return annotations.count()
-
-    @property
-    def dataset(self):
-        if self._dataset is None:
-            self._dataset = DatasetModel.objects(id=self.dataset_id).first()
-        return self._dataset
-
-    def __call__(self):
-
-        image = im.Image.from_path(self.path)
-        for annotation in AnnotationModel.objects(image_id=self.id, deleted=False).all():
-            if not annotation.is_empty():
-                image.add(annotation())
-
-        return image
-    
-    def can_delete(self, user):
-        return user.can_delete(self.dataset)
-    
-    def can_download(self, user):
-        return user.can_download(self.dataset)
-    
-    # TODO: Fix why using the functions throws an error
-    def permissions(self, user):
-        return {
-            'delete': True,
-            'download': True
-        }
-    
-    def add_event(self, e):
-        event = Event()
-        event.set_data(e)
-        self.update(push__events=event)
-
-
-class AnnotationModel(db.DynamicDocument):
-
-    id = db.SequenceField(primary_key=True)
-    image_id = db.IntField(required=True)
-    category_id = db.IntField(required=True)
-    dataset_id = db.IntField()
-
-    segmentation = db.ListField(default=[])
-    area = db.IntField(default=0)
-    bbox = db.ListField(default=[0, 0, 0, 0])
-    iscrowd = db.BooleanField(default=False)
-
-    creator = db.StringField(required=True)
-    width = db.IntField()
-    height = db.IntField()
-
-    color = db.StringField()
-    
-    keypoints = db.ListField(default=[])
-
-    metadata = db.DictField(default={})
-    paper_object = db.ListField(default=[])
-
-    deleted = db.BooleanField(default=False)
-    deleted_date = db.DateTimeField()
-
-    def __init__(self, image_id=None, **data):
-
-        image = ImageModel.objects(id=image_id).first()
-
-        if image is not None:
-            data['image_id'] = image_id
-            data['width'] = image.width
-            data['height'] = image.height
-            data['dataset_id'] = image.dataset_id
-
-        super(AnnotationModel, self).__init__(**data)
-
-    def save(self, copy=False, *args, **kwargs):
-
-        if not self.dataset_id and not copy:
-            dataset = DatasetModel.objects(id=self.dataset_id).first()
-
-            if dataset is not None:
-                self.metadata = dataset.default_annotation_metadata.copy()
-
-        if self.color is None:
-            self.color = im.Color.random().hex
-
-        if current_user:
-            self.creator = current_user.username
-        else:
-            self.creator = 'system'
-
-        return super(AnnotationModel, self).save(*args, **kwargs)
-
-    def is_empty(self):
-        return len(self.segmentation) == 0 or self.area == 0
-
-    def mask(self):
-        """ Returns binary mask of annotation """
-        mask = np.zeros((self.height, self.width))
-        pts = [
-            np.array(anno).reshape(-1, 2).round().astype(int)
-            for anno in self.segmentation
-        ]
-        mask = cv2.fillPoly(mask, pts, 1)
-        return mask
-
-    def clone(self):
-        """ Creates a clone """
-        create = json.loads(self.to_json())
-        del create['_id']
-
-        return AnnotationModel(**create)
-
-    def __call__(self):
-
-        category = CategoryModel.objects(id=self.category_id).first()
-        if category:
-            category = category()
-
-        data = {
-            'image': None,
-            'category': category,
-            'color': self.color,
-            'polygons': self.segmentation,
-            'width': self.width,
-            'height': self.height,
-            'metadata': self.metadata
-        }
-
-        return im.Annotation(**data)
-
-
-class CategoryModel(db.DynamicDocument):
-
-    id = db.SequenceField(primary_key=True)
-    name = db.StringField(required=True, unique_with=['creator'])
-    supercategory = db.StringField(default="")
-    color = db.StringField(default=None)
-    metadata = db.DictField(default={})
-
-    creator = db.StringField(default="unknown")
-    deleted = db.BooleanField(default=False)
-    deleted_date = db.DateTimeField()
-
-    keypoint_edges = db.ListField(default=[])
-    keypoint_labels = db.ListField(default=[])
-    
-
-    @classmethod
-    def bulk_create(cls, categories):
-
-        if not categories:
-            return []
-
-        category_ids = []
-        for category in categories:
-            category_model = CategoryModel.objects(name=category).first()
-
-            if category_model is None:
-                new_category = CategoryModel(name=category)
-                new_category.save()
-                category_ids.append(new_category.id)
-            else:
-                category_ids.append(category_model.id)
-
-        return category_ids
-
-    def save(self, *args, **kwargs):
-
-        if not self.color:
-            self.color = im.Color.random().hex
-
-        if current_user:
-            self.creator = current_user.username
-        else:
-            self.creator = 'system'
-      
-        return super(CategoryModel, self).save(*args, **kwargs)
-
-    def __call__(self):
-        """ Generates imantics category object """
-        data = {
-            'name': self.name,
-            'color': self.color,
-            'parent': self.supercategory,
-            'metadata': self.metadata,
-            'id': self.id
-        }
-        return im.Category(**data)
-    
-    def is_owner(self, user):
-
-        if user.is_admin:
-            return True
-        
-        return user.username.lower() == self.creator.lower()
-    
-    def can_edit(self, user):
-        return self.is_owner(user)
-    
-    def can_delete(self, user):
-        return self.is_owner(user)
-
-
-class LicenseModel(db.DynamicDocument):
-    id = db.SequenceField(primary_key=True)
-    name = db.StringField()
-    url = db.StringField()
-
-
-class TaskModel(db.DynamicDocument):
-    id = db.SequenceField(primary_key=True)
-    
-    # Type of task: Importer, Exporter, Scanner, etc.
-    group = db.StringField(required=True)
-    name = db.StringField(required=True) 
-    desciption = db.StringField()
-
-    creator = db.StringField()
-
-    #: Start date of the executor 
-    start_date = db.DateTimeField()
-    #: End date of the executor 
-    end_date = db.DateTimeField()
-    completed = db.BooleanField(default=False)
-    failed = db.BooleanField(default=False)
-    has_download = db.BooleanField(default=False)
-    
-    # If any of the information is relevant to the task
-    # it should be added
-    dataset_id = db.IntField()
-    image_id = db.IntField()
-    category_id = db.IntField()
-
-    progress = db.FloatField(default=0, min_value=0, max_value=100)
-
-    logs = db.ListField(default=[])
-    errors = db.IntField(default=0)
-    warnings = db.IntField(default=0)
-
-    priority = db.IntField()
-
-    metadata = db.DictField(default={})
-
-    _update_every = 10
-    _progress_update = 0
-
-    def error(self, string):
-        self._log(string, level="ERROR")
-    
-    def warning(self, string):
-        self._log(string, level="WARNING")
-    
-    def info(self, string):
-        self._log(string, level="INFO")
-    
-    def _log(self, string, level):
-
-        level = level.upper()
-        date = datetime.datetime.now().strftime("%d-%m-%Y %H:%M:%S")
-        
-        message = f"[{date}] [{level}] {string}"
-
-        statment = {
-            'push__logs': message
-        }
-
-        if level == "ERROR":
-            statment['inc__errors'] = 1
-            self.errors += 1
-        
-        if level == "WARNING":
-            statment['inc__warnings'] = 1
-            self.warnings += 1
-
-        self.update(**statment)
-
-    def set_progress(self, percent, socket=None):
-
-        self.update(progress=int(percent), completed=(percent >= 100))
-
-        # Send socket update every 10%
-        if self._progress_update < percent or percent >= 100:
-            
-            if socket is not None:
-                logger.debug(f"Emitting {percent} progress update for task {self.id}")
-
-                socket.emit('taskProgress', {
-                    'id': self.id,
-                    'progress': percent,
-                    'errors': self.errors,
-                    'warnings': self.warnings
-                }, broadcast=True)
-            
-            self._progress_update += self._update_every
-
-    def start(self, target, *args, **kwargs):
-        
-        from .sockets import socketio
-
-        thread = socketio.start_background_task(
-            target,
-            task=self,
-            socket=socketio,
-            *args,
-            **kwargs
-        )
-        return thread
-    
-    def api_json(self):
-        return {
-            "id": self.id,
-            "name": self.name
-        }
-
-
-class UserModel(db.DynamicDocument, UserMixin):
-    password = db.StringField(required=True)
-    username = db.StringField(max_length=25, required=True, unique=True)
-    email = db.StringField(max_length=30)
-
-    name = db.StringField()
-    last_seen = db.DateTimeField()
-
-    is_admin = db.BooleanField(default=False)
-
-    preferences = db.DictField(default={})
-    permissions = db.ListField(defualt=[])
-
-    def save(self, *args, **kwargs):
-
-        self.last_seen = datetime.datetime.now()
-
-        return super(UserModel, self).save(*args, **kwargs)
-
-    @property
-    def datasets(self):
-        self._update_last_seen()
-
-        if self.is_admin:
-            return DatasetModel.objects
-
-        return DatasetModel.objects(Q(owner=self.username) | Q(users__contains=self.username))
-
-    @property
-    def categories(self):
-        self._update_last_seen()
-
-        if self.is_admin:
-            return CategoryModel.objects
-
-        dataset_ids = self.datasets.distinct('categories')
-        return CategoryModel.objects(Q(id__in=dataset_ids) | Q(creator=self.username))
-
-    @property
-    def images(self):
-        self._update_last_seen()
-
-        if self.is_admin:
-            return ImageModel.objects
-
-        dataset_ids = self.datasets.distinct('id')
-        return ImageModel.objects(dataset_id__in=dataset_ids)
-
-    @property
-    def annotations(self):
-        self._update_last_seen()
-
-        if self.is_admin:
-            return AnnotationModel.objects
-
-        image_ids = self.images.distinct('id')
-        return AnnotationModel.objects(image_id__in=image_ids)
-
-    def can_view(self, model):
-        if model is None:
-            return False
-
-        return model.can_view(self)
-    
-    def can_download(self, model):
-        if model is None:
-            return False
-
-        return model.can_download(self)
-        
-    def can_delete(self, model):
-        if model is None:
-            return False
-        return model.can_delete(self)
-
-    def can_edit(self, model):
-        if model is None:
-            return False
-
-        return model.can_edit(self)
-
-    def _update_last_seen(self):
-        self.update(last_seen=datetime.datetime.now())
-
-
-class SessionEvent(db.EmbeddedDocument):
-    NAME = "SESSION"
-
-    user = db.StringField(required=True)
-    milliseconds = db.IntField(default=0, min_value=0)
-    tools_used = db.ListField(default=[])
-
-    @classmethod
-    def create(self, start, user=None, end=None, tools=[]):
-        
-        if user is None:
-            user = current_user.username
-
-        if end is None:
-            end = time.time()
-
-        return SessionEvent(
-            user=user,
-            milliseconds=int((end-start)*1000)
-        )
-
-
-# https://github.com/MongoEngine/mongoengine/issues/1171
-# Use this methods until a solution is found
-def upsert(model, query=None, update=None):
-
-    if not update:
-        update = query
-
-    if not query:
-        return None
-
-    found = model.objects(**query)
-
-    if found.first():
-        return found.modify(new=True, **update)
-
-    new_model = model(**update)
-    new_model.save()
-
-    return new_model
-
-
-def create_from_json(json_file):
-
-    with open(json_file) as file:
-
-        data_json = json.load(file)
-        for category in data_json.get('categories', []):
-            name = category.get('name')
-            if name is not None:
-                upsert(CategoryModel, query={"name": name}, update=category)
-
-        for dataset_json in data_json.get('datasets', []):
-            name = dataset_json.get('name')
-            if name:
-                # map category names to ids; create as needed
-                category_ids = []
-                for category in dataset_json.get('categories', []):
-                    category_obj = {"name": category}
-
-                    category_model = upsert(CategoryModel, query=category_obj)
-                    category_ids.append(category_model.id)
-
-                dataset_json['categories'] = category_ids
-                upsert(DatasetModel, query={ "name": name}, update=dataset_json)
-
