+++ docker-compose.dev.yml
@@ -28,12 +28,12 @@ services:
     ports:
       - "5000:5000"
     volumes:
-      - "./datasets:/datasets"
       - "./:/workspace/"
+      - "./datasets:/datasets"
       - "./models:/models"
     environment:
       - FILE_WATCHER=true
-      - FLASK_APP=app
+      - FLASK_APP=webserver
       - FLASK_ENV=development
       # - MASK_RCNN_FILE=/models/mask_rcnn_coco.h5
       # # Default coco classes (order matters)
