+++ libs/database/lisence.py
@@ -1,0 +1,10 @@
+from mongoengine import *
+
+
+class LicenseModel(DynamicDocument):
+    id = SequenceField(primary_key=True)
+    name = StringField()
+    url = StringField()
+
+
+__all__ = ["LicenseModel"]
