+++ webserver/api/images.py
@@ -4,8 +4,14 @@ from werkzeug.datastructures import FileStorage
 from flask import send_file
 
 from ..util import query_util, coco_util
-from ..models import *
+from database import (
+    ImageModel,
+    DatasetModel,
+    CategoryModel,
+    AnnotationModel
+)
 
+from PIL import Image
 import datetime
 import os
 import io
