+++ webserver/api/info.py
@@ -1,8 +1,8 @@
 from flask_restplus import Namespace, Resource, reqparse
 
-from ..config import Config
+from config import Config
 from ..util.version_util import get_tag
-from ..models import UserModel
+from database import UserModel
 
 
 api = Namespace('info', description='Software related operations')
@@ -19,7 +19,7 @@ class Info(Resource):
             "demo": "https://annotator.justinbrooks.ca/",
             "repo": "https://github.com/jsbroks/coco-annotator",
             "git": {
-                "tag": get_tag()
+                "tag": Config.VERSION
             },
             "login_enabled": not Config.LOGIN_DISABLED,
             "total_users": UserModel.objects.count(),
